Add ActiveSession model to use it for security
We're currently setting the user's ID in the session. Even though that value is encrypted, the encrypted value doesn't change since it's based on the user id which doesn't change. This means that if a bad actor were to get a copy of the session they would have access to a victim's account in perpetuity
Ref: https://stevepolito.design/blog/rails-authentication-from-scratch/#step-18-store-session-in-the-database